requirements.txt
# Music Streaming Analytics Platform
# Python Dependencies

# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Database
psycopg2-binary>=2.9.6
sqlalchemy>=2.0.0

# Machine Learning
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Spotify API
spotipy>=2.23.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Data Validation
pandera>=0.16.0

# Logging
loguru>=0.7.0

# Progress Bars
tqdm>=4.65.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0

# HTTP Requests
requests>=2.31.0

# Date/Time handling
python-dateutil>=2.8.2
